#747: Reordered typealiases to see them better
