+++ AtlasSDK/AtlasSDK/Endpoints/GetGuestCheckoutEndpoint.swift
@@ -2,8 +2,6 @@
 //  Copyright © 2016-2017 Zalando SE. All rights reserved.
 //
 
-public typealias CheckoutToken = String
-
 struct GetGuestCheckoutEndpoint: CheckoutGatewayEndpoint {
 
     let config: Config
