+++ AtlasSDK/AtlasSDK/Models/CheckoutAddresses.swift
@@ -4,9 +4,6 @@
 
 import Foundation
 
-public typealias BillingAddress = EquatableAddress
-public typealias ShippingAddress = EquatableAddress
-
 public struct CheckoutAddresses {
 
     public let billingAddress: BillingAddress?
