+++ AtlasSDK/AtlasSDK/Types.swift
@@ -1,0 +1,19 @@
+//
+//  Copyright © 2017 Zalando SE. All rights reserved.
+//
+
+import Foundation
+
+public typealias SKU = String
+
+public typealias CartId = String
+public typealias CheckoutId = String
+public typealias CheckoutToken = String
+public typealias CustomerNumber = String
+
+public typealias AddressId = String
+public typealias BillingAddress = EquatableAddress
+public typealias ShippingAddress = EquatableAddress
+
+public typealias MoneyAmount = Decimal
+public typealias Currency = String
